Add done functionality to TodoListItem
This will refactoring AddTodoItem functionality to TodoList and
TodoListItem components. The main reason for this is that clicking on
checkbox input and making a todo done is bargin from AddTodoItem so I
just used it for changing state and responsibility of
showing/completeting todos are belong to TodoList and TodoListItem.